requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
tensorflow==2.16.1
numpy==1.23.5
pandas==1.5.3
requests==2.31.0
scikit-learn==1.2.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
